Update FAQ page to use CSF
