+++ packages/@codotype/ui/src/pages/faq/component.tsx
